tests/requirements.txt
pytest
pytest-xdist
allure-pytest
docker
retry
multipledispatch
requests_toolbelt